Add files via upload
Create the quiz_game_page.html file for the game page.
So start coding in quiz_game_login.js file -
Add the addUser() function which is been called onclick of the login button from the login page.
So addUser() function takes value from the input boxes of player 1 and player 2 and add’s it in the local storage, and redirects to the quiz_game_page.html. This is already done in class 90 so it should not be difficult for you to do it.
If you get stuck somewhere, you can refer to the game_login.js file from class 90, because almost everything is the same.
After completing JS code for the game login page, start writing HTML code for quiz_game_page.html, which you just created.
This HTML page should have -
1. Bootstrap links
2. Your style sheet link, which you had created in the previous project. Like this


Your javascript link (make a blank JS file and put the name of the file in this HTML file, we will do JS coding of this file in the next class), something like this -

Before the body ends
Now add HTML elements with id for holding the player 1 and player 2 name and their scores. Like this -


Create a div with the bootstrap class container, and create a center tag, and put all the HTML elements inside this.

Now add heading tag which is holding the heading of the game, this -

like this 
Then add heading elements with id, which will be used to hold the name of questioner and answer turn
Then a div with id=’output’, which will be used to store the question, input box for the answer and the check button, this -

CSS for the above div


Then two headings and two input boxes for the numbers, like this - 

And in last there should be a send button, and give an on click function send() to it, this send() function we will write in the next class


For CSS you can refer to the game.css which has been downloaded by you in class 89.
If you get stuck somewhere, you can refer to the game_page.html file which has been downloaded by you in class 90, because almost everything is the same.

Use you creativity and make this page look good.
Style it your own way. Only remember it should have all the HTML elements as mentioned above

HINT - - Use bootstrap classes for making it responsive and look beautiful.



Submitting the Project:- If you have created project-math-quiz folder on github then -

Reupload all the files on which you have worked in the current project, in the project-math-quiz folder(this folder has been created by you in the previous project) on GitHub. You can get the steps to do this by clicking on thislink

- If you have not created project-math-quiz folder on github then -

Create a folder project-C89 and upload all the files related to the project-math-quiz project on GitHub. You can get the steps to do this by clicking on thislink.

Copy the hosted link which you will get after uploading all your files on GitHub and submit it in the Student Dashboard Projects panel against the correct class number.
Start Your Project

Read all the instructions carefully to understand the project

Click on "Start Project" to go to the project playground and start your project
Start Project
Submit Your Project
Learn how to submit your projectvideo
Paste your project URL
Submit Project
Class Summary
This project is based on your last class ADV-C90